feat: add current work
Leaving off at not being able to parse keys that have no values but have
a comment at the end of the line. For some reason this is breaking the
parser. The comment shouldn't make it into the table but it is. Need to
figure out why it's making it into the table or just strip it if it's
present.